Include message offset in yielded data
Add (at a configurable key) the source of a message as
"<topic>-<partition>-<offset>" to the hash of data before yielding it.

Example usage: In Wally, we put this value in a uniquely-indexed field
on any data we create. This prevents duplicate records in the (unlikely)
event that Wally sees the same message twice.