move KafkaOffset into this gem as a base module
Pursuing the dont-write-records-with-invalid-current-offsets issue, it
seems like it would be nicer if the code shared by multiple clients of
this gem (akhd which needs to be patched) was shared in the gem, rather
than than being duplicated everywhere.

This is a slight challenge, since the implementation depends on
minidoc, which can't easily be specifed as a gem dep since it's a github
project that hasn't actually been released as a gem (which means it
can't really be expressed in the gemspec).

This is a compromise: we move the implementation into a base module that
client code can easily include. So the implementation lives here, but
minidoc is not a direct gem dependency.

This also implements part of the fix we need for the original current
offset issue, which is: don't overwrite #current on the offset class.