require kafka to acknowledge writes
We theorize that our kafka crash earlier today was made worse by not
requiring kafka to acknowledge & agree on writes.

This setting is the strongest level of durability: kafka doesn't
acknowledge the the message until all replicas are in sync. (Kafka docs:
https://kafka.apache.org/08/configuration.html).

This may not be what we want (for example, I'm not sure what happens with this setting when a
kafka node goes down: does the cluster acknowledge writes when all
remaining nodes agree, or does it just stop acknowledging altogether?).
But I'm pretty sure we don't want the default either.