Remove transaction around consumer loop
We have a case where one system creates data then sends a message to
another system that reads it. Since both actions occur in one of these
transaction loops, the second system cannot see the data created by the
first (if it hasn't yet committed its transaction).

Provided there are no downsides, removing the transaction(s) should be
an overall simplification and resolve the issue.

The original benefits of the transaction are tenuous at best. Because
the offset is being moved before the transaction is opened, this change
doesn't change the retry semantics in the case of errors (those retry
semantics may not be what they want, but this change at least preserves
them for now). This change would make it possible for partial updates
being made within the processing loop to not get entirely rolled back in
the case of errors. But our systems need to be robust to this anyway and
should not rely on multi-document transactions in general.